+++ packages/app-builder-lib/src/targets/appimage/appImageUtil.ts
@@ -1,13 +1,14 @@
 import { Arch, copyDir, copyFile, exec, exists, InvalidConfigurationError, log } from "builder-util"
-import fs from "fs-extra"
+import * as fs from "fs-extra"
 import * as path from "path"
-import { FileAssociation } from "../../options/FileAssociation.js"
-import { IconInfo } from "../../platformPackager.js"
-import { getAppImageTools } from "../../toolsets/linux.js"
-import { copyIcons, copyMimeTypes } from "./appLauncher.js"
-import { appendBlockmap } from "../differentialUpdateInfoBuilder.js"
+import { FileAssociation } from "../../options/FileAssociation"
+import { IconInfo } from "../../platformPackager"
+import { getAppImageTools } from "../../toolsets/linux"
+import { copyIcons, copyMimeTypes } from "./appLauncher"
+import { appendBlockmap } from "../differentialUpdateInfoBuilder"
 import { BlockMapDataHolder } from "builder-util-runtime"
-import { APP_RUN_ENTRYPOINT } from "./AppImageTarget.js"
+import { APP_RUN_ENTRYPOINT } from "./AppImageTarget"
+import { ToolsetConfig } from "app-builder-lib/src/configuration"
 
 interface Options {
   productName: string
@@ -28,7 +29,7 @@ export interface AppImageBuilderOptions {
   options: Options
 }
 
-export async function buildAppImage(opts: AppImageBuilderOptions): Promise<BlockMapDataHolder> {
+export async function buildAppImage(appimageToolVersion: ToolsetConfig["appimage"], opts: AppImageBuilderOptions): Promise<BlockMapDataHolder> {
   const { stageDir, output, appDir, options, arch } = opts
 
   try {
@@ -37,7 +38,7 @@ export async function buildAppImage(opts: AppImageBuilderOptions): Promise<Block
     // Write AppRun launcher and related files
     await writeAppLauncherAndRelatedFiles(opts)
 
-    const { runtimeLibraries: libraries, runtime, mksquashfs } = await getAppImageTools(arch)
+    const { runtimeLibraries: libraries, runtime, mksquashfs } = await getAppImageTools(appimageToolVersion, arch)
     await copyDir(libraries, path.join(stageDir, "usr", "lib"))
 
     // Copy app directory to stage
@@ -201,7 +202,7 @@ function hasEula(config: AppRunScript): config is AppRunScriptWithEula {
 function generateAppRunScript(config: AppRunScript): string {
   const eulaEnabled = hasEula(config)
 
-  return `#!/bin/bash
+  return `#!/usr/bin/env bash
 set -e
 
 THIS="$0"