+++ test/src/HoistedNodeModuleTest.ts
@@ -98,12 +98,10 @@ describe.ifNotWindows("node_module collectors", () => {
         projectDirCreated: async (projectDir, _tmpDir, testEnv) => {
           await modifyPackageJson(projectDir, data => {
             data.dependencies = {
-              "electron-updater": "6.8.3",
-              express: "4.0.0",
+              "is-odd": "3.0.1",
             }
             data.devDependencies = {
-              electron: "23.2.0",
-              "electron-builder": "26.8.1",
+              electron: ELECTRON_VERSION,
               "fs-extra": "11.0.0",
               "patch-package": "8.0.0",
             }
@@ -151,13 +149,10 @@ describe.ifNotWindows("node_module collectors", () => {
         projectDirCreated: async (projectDir, _tmpDir, testEnv) => {
           await modifyPackageJson(projectDir, data => {
             data.dependencies = {
-              "electron-updater": "6.8.3",
-              express: "4.0.0",
-              "patch-package": "8.0.0",
+              "is-odd": "3.0.1",
             }
             data.devDependencies = {
-              electron: "23.2.0",
-              "electron-builder": "26.8.1",
+              electron: ELECTRON_VERSION,
               "fs-extra": "11.0.0",
             }
             data.build.directories = {
@@ -476,7 +471,7 @@ describe.ifNotWindows("node_module collectors", () => {
               "electron-clear-data": "1.0.5",
             }
             data.devDependencies = {
-              electron: "34.0.2",
+              electron: ELECTRON_VERSION,
             }
           })
           await spawn("yarn", ["install"], {