+++ packages/electron-updater/CHANGELOG.md
@@ -1,5 +1,21 @@
 ## 4.3.0
 
+## 6.8.4
+
+### Patch Changes
+
+<details><summary>Updated 1 dependency</summary>
+
+<small>
+
+[`b7df0bc`](https://github.com/electron-userland/electron-builder/commit/b7df0bcef5790814fd6df187346dcec625dc681e)
+
+</small>
+
+- `builder-util-runtime@9.6.0`
+
+</details>
+
 ## 6.8.3
 
 ### Patch Changes